requirements.txt
flask
flask-cors
PyJWT
redis